aep-sdk/pyproject.toml
[tool.poetry]
name = "aep"
version = "0.1.0"
description = "Attention Extended Payloads (AEP) SDK for logging and utilizing interaction focus."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
langchain-core = "*"
msgpack = "*"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
openai = "1.23.6"
httpx = "0.27.2"
langchain = "0.1.20"
langchain-openai = "0.1.5"
faiss-cpu = "1.7.4"
langgraph = "0.0.40"
unstructured = "^0.14"
langchainhub = "^0.1"
markdown = "^3.6"
docarray = "^0.40"
nltk = "^3.8"
pandas = "^2.2"
matplotlib = "^3.8"
seaborn = "^0.13"
pyyaml = "^6.0"
portalocker = "^3.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
# Add other dev dependencies like black, ruff if needed

[tool.poetry.scripts]
aep = "aep.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 